+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Format named arguments anywhere (#1072).
 * Format enhanced enums (#1075).
+* Format "super." parameters (#1091).
 
 # 2.2.1